edit tickets

+++ src/components/queue/QueueListItem/QueueListItem.tsx
@@ -1,4 +1,4 @@
-import React, {FC} from "react";
+import React, {FC, useState} from "react";
 import {Box, Paper, Stack, Typography} from "@mui/material";
 import IconButton from "@components/shared/IconButton";
 import DeleteIcon from "@mui/icons-material/Delete";
@@ -6,8 +6,11 @@ import EditIcon from "@mui/icons-material/Edit";
 import ConfirmationNumberIcon from '@mui/icons-material/ConfirmationNumber';
 import ConfirmationNumberOutlinedIcon from '@mui/icons-material/ConfirmationNumberOutlined';
 import CheckBoxIcon from '@mui/icons-material/CheckBox';
-import QueueAPI, {Ticket, Queue, TicketStatus} from "@util/queue/api";
+import HelpCenterIcon from '@mui/icons-material/HelpCenter';
+import QueueAPI, {Ticket, TicketStatus} from "@util/queue/api";
 import {toast} from "react-hot-toast";
+import {useAuth} from "@util/auth/hooks";
+import EditTicketDialog from "@components/queue/EditTicketDialog";
 
 export interface QueueListItemProps {
     queueID: string;
@@ -15,38 +18,52 @@ export interface QueueListItemProps {
 }
 
 const QueueListItem: FC<QueueListItemProps> = ({queueID, ticket}) => {
+    const {currentUser, isAuthenticated} = useAuth();
+    const [editTicketDialog, setEditTicketDialog] = useState(false);
 
     const claimed = ticket.status === TicketStatus.StatusClaimed;
-    
-    return (
+    const missing = ticket.status === TicketStatus.StatusMissing;
+    const isTA = currentUser && (currentUser.coursePermissions[queueID] != undefined);
+    const ownsTicket = currentUser && ticket.createdBy.Email === currentUser.email;
+
+    const staffPerm = isTA || currentUser?.isAdmin;
+    const ownedPerm = staffPerm || ownsTicket;
+
+    // TODO: EditTicketDialog should be in QueueList to avoid rendering many times.
+
+    return (<>
+        <EditTicketDialog open={editTicketDialog} onClose={() => setEditTicketDialog(false)} id={ticket.id} queueID={queueID as string} status={ticket.status} />
         <Paper variant="outlined">
             <Box p={2.5}>
                 <Stack direction="row" justifyContent="space-between" alignItems="center" spacing={2}>
                     <Box>
-                        <Typography fontSize={16} fontWeight={600}>{ticket.createdBy.DisplayName}</Typography>
+                        <Typography fontSize={16} fontWeight={600}>{ticket.createdBy.DisplayName}{missing && " (MISSING)"}</Typography>
                         <Typography fontSize={14}>{ticket.description}</Typography>
                     </Box>
                     <Box>
-                        {!claimed && <IconButton label="Claim ticket" edge="end" aria-label="delete" onClick={() => QueueAPI.editTicket(ticket.id, queueID, TicketStatus.StatusClaimed, ticket.description)}>
+                        {staffPerm && !claimed && <IconButton label="Claim ticket" edge="end" aria-label="claim" onClick={() => QueueAPI.editTicket(ticket.id, queueID, TicketStatus.StatusClaimed, ticket.description)}>
                             <ConfirmationNumberOutlinedIcon/>
                         </IconButton>}
-                        {claimed && <IconButton label="Checkoff ticket" edge="end" aria-label="delete" onClick={() => QueueAPI.editTicket(ticket.id, queueID, TicketStatus.StatusComplete, ticket.description)}>
+                        {staffPerm && claimed && <IconButton label="Mark as missing" edge="end" aria-label="mark-missing" onClick={() => QueueAPI.editTicket(ticket.id, queueID, TicketStatus.StatusMissing, ticket.description)}>
+                            <HelpCenterIcon/>
+                        </IconButton>}
+                        {staffPerm && claimed && <IconButton label="Checkoff ticket" edge="end" aria-label="checkoff" onClick={() => QueueAPI.editTicket(ticket.id, queueID, TicketStatus.StatusComplete, ticket.description)}>
                             <CheckBoxIcon/>
                         </IconButton>}
-                        {claimed && <IconButton label="Unclaim ticket" edge="end" aria-label="delete" onClick={() => QueueAPI.editTicket(ticket.id, queueID, TicketStatus.StatusWaiting, ticket.description)}>
+                        {staffPerm && claimed && <IconButton label="Unclaim ticket" edge="end" aria-label="unclaim" onClick={() => QueueAPI.editTicket(ticket.id, queueID, TicketStatus.StatusWaiting, ticket.description)}>
                             <ConfirmationNumberIcon/>
                         </IconButton>}
-                        <IconButton label="Edit ticket" edge="end" aria-label="delete" onClick={() => toast.error(`Feature unimplemented.`)}>
+                        {ownedPerm && <IconButton label="Edit ticket" edge="end" aria-label="edit" onClick={() => setEditTicketDialog(true)}>
                             <EditIcon/>
-                        </IconButton>
-                        <IconButton label="Delete ticket" edge="end" aria-label="delete" onClick={() => QueueAPI.deleteTicket(ticket.id, queueID)}>
+                        </IconButton>}
+                        {ownedPerm && <IconButton label="Delete ticket" edge="end" aria-label="delete" onClick={() => QueueAPI.deleteTicket(ticket.id, queueID)}>
                             <DeleteIcon/>
-                        </IconButton>
+                        </IconButton>}
                     </Box>
                 </Stack>
             </Box>
         </Paper>
-    );
+    </>);
 };
 
 export default QueueListItem;
